% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Gaussian_corrupted_1D.R
\name{data_gaussian_corrupted_1D}
\alias{data_gaussian_corrupted_1D}
\title{Generate Corrupted Gaussian Data (Excluding Specified Ranges)}
\usage{
data_gaussian_corrupted_1D(
  N = 1000,
  m = 0,
  s2 = 1,
  l = c(-1.5, 1.3),
  r = c(-1, 2)
)
}
\arguments{
\item{N}{Integer, the number of points to be generated.}

\item{m}{Numeric, the mean of the Gaussian distribution.}

\item{s2}{Numeric, the variance of the Gaussian distribution.}

\item{l}{Numeric vector, the lower bounds of the ranges to exclude.}

\item{r}{Numeric vector, the upper bounds of the ranges to exclude.}
}
\value{
A numeric vector of length `N` containing the corrupted Gaussian data,
        with points removed from the specified ranges.
}
\description{
This function generates `N` points from a Gaussian distribution with mean `m` and variance `s2`,
excluding any points that fall within specified ranges (given by `l` and `r`).
}
\details{
This function generates `N` points from a Gaussian distribution. If any point falls within the
ranges defined by the vectors `l` and `r`, it is excluded from the result.
The points that remain will be outside the specified ranges.
}
\examples{
# Generate 1000 corrupted Gaussian points with mean 0 and variance 1,
# excluding points in the ranges [-1.5, -1] and [1.3, 2].
data_gaussian_corrupted_1D(N = 1000, m = 0, s2 = 1, l = c(-1.5, 1.3), r = c(-1, 2))

}
\seealso{
\code{\link{rnorm}}
}
