% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Gaussian_mixture_1D.R
\name{gaussian_mixture_1D}
\alias{gaussian_mixture_1D}
\title{Generate Data from a 1D Gaussian Mixture Model}
\usage{
gaussian_mixture_1D(N, means = c(0, 2), sds = c(2, 3), weights = c(0.4, 0.6))
}
\arguments{
\item{N}{Integer, total number of data points to generate.}

\item{means}{Numeric vector of length `k`, specifying the means of the `k` Gaussians.}

\item{sds}{Numeric vector of length `k`, specifying the standard deviations of the `k` Gaussians.}

\item{weights}{Numeric vector of length `k`, specifying the mixture weights (should sum to 1).}
}
\value{
A numeric vector of `N` values drawn from the specified Gaussian mixture.
}
\description{
This function generates `N` data points from a mixture of `k` Gaussian distributions
with given means, standard deviations, and mixture weights.
}
\details{
This function randomly assigns each data point to one of the `k` Gaussians based on the
given mixture weights and then samples from the corresponding normal distribution.
}
\examples{
# Generate 1000 points from a mixture of two Gaussians
data <- gaussian_mixture_1D(N = 1000, means = c(-2, 3), sds = c(1, 2), weights = c(0.4, 0.6))
hist(data, breaks = 50, probability = TRUE, col = "lightblue", main = "Gaussian Mixture")

}
