% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecc.R
\name{ecc1D}
\alias{ecc1D}
\title{Plot Euler Characteristic Curve for 1D Data}
\usage{
ecc1D(data)
}
\arguments{
\item{data}{Numeric vector, a set of data points in 1D.}
}
\value{
A plot of the Euler characteristic curve with the radius (merging time) on the x-axis
        and the Euler characteristic on the y-axis.
}
\description{
This function computes and plots the Euler characteristic curve for a given set of 1D data points.
The Euler characteristic is calculated by measuring the number of connected components minus the number of holes
at different radius levels based on the distances between consecutive points.
}
\details{
The function first computes the distances between consecutive sorted data points, and then uses these distances
to create the Euler characteristic curve. The Euler characteristic is calculated as the difference between the number
of connected components and the number of holes as the radius (or merging time) increases.
The curve is plotted on a logarithmic scale for both the x-axis (radius) and y-axis (Euler characteristic).
}
